Refactor: 서버 실행 파일 변경
