+++ package.json
@@ -3,8 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "cross-env NODE_ENV=production ts-node ./bin/www.ts",
-    "dev": "cross-env NODE_ENV=develop nodemon --exec ts-node ./bin/www.ts"
+    "start": "cross-env NODE_ENV=production ts-node ./src/server.ts",
+    "dev": "cross-env NODE_ENV=development nodemon --exec ts-node ./src/server.ts"
   },
   "dependencies": {
     "@types/cookie-parser": "^1.4.2",
@@ -16,6 +16,7 @@
     "cors": "^2.8.5",
     "debug": "~2.6.9",
     "dotenv": "^8.2.0",
+    "errorhandler": "^1.5.1",
     "express": "^4.16.4",
     "http-errors": "~1.6.3",
     "jade": "~1.11.0",
@@ -25,6 +26,7 @@
     "swagger-ui-express": "^4.1.6"
   },
   "devDependencies": {
+    "@types/errorhandler": "^1.5.0",
     "@types/express": "^4.17.9",
     "@types/node": "^14.14.16",
     "cross-env": "^7.0.3",
