+++ src/app.ts
@@ -8,16 +8,14 @@ import cors from "cors";
 import dotenv from "dotenv";
 
 import indexRouter from "./routes/index";
+import { normalizePort } from "./middleWares/index";
 
 const app: express.Application = express();
 
 // use env_values
 dotenv.config();
 
-// view engine setup
-app.set("views", path.join(__dirname, "views"));
-app.set("view engine", "jade");
-
+app.set("port", normalizePort(process.env.PORT || "3000"));
 app.use(cors());
 if (process.env.NODE_ENV === "production") {
   app.use(logger("combined"));
@@ -36,12 +34,12 @@ app.use(bodyParser.json());
 app.use("/", indexRouter);
 
 // catch 404 and forward to error handler
-app.use(function (req: Request, res: Response, next) {
+app.use((req: Request, res: Response, next: NextFunction) => {
   next(createError(404));
 });
 
 // error handler
-app.use(function (err: any, req: Request, res: Response, next: NextFunction) {
+app.use((err: any, req: Request, res: Response, next: NextFunction) => {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get("env") === "development" ? err : {};
